Using AverageIfs in Excel to calculate the days of week
Essentially, how the averages for the specific days of the week for both January and February are calculated are by using the AverageIfs function. As an example, to calculate the average revenue made on Mondays, the AverageIf function will take in the values starting and including G4 and ending and including G63. This argument in the function represents the data that is used to find the averages. I could have selected just the Mondays from the 2 months, January and February, but that isn't dynamic, as when March , April and May are added to the excel workbook I would have to add in over 12 more values, the cells to perform the =AVERAGE() function. With AverageIfs function, the benefit is that you can narrow and refine your search of a data set based on criteria in a wholly different column or row. So, in this example the data we are finding the averages form is of course the revenue starting at G4 and ending at G63, but we don't want to have to manually select all of the Mondays that are included in the data. That is where the second and third arguments of the AverageIfs function come to the rescue. The second argument is E4:E63, which is all the days of the week repeated, which of course we only want to find only the specific day of the week for calculating the averages. This is where E4 comes in, which is an element that contains the search criteria "Monday", which will be selected from all of the days. Now, the only problem that arises is that obviously, if you divide the 2 months by all of the Mondays, the answer won't always be in proper dollars and cents format. You can have infinitely repeating digits that are fractions of a penny which are meaningless to most companies. This is why the ROUND function is used to clean up the mess that the calculation makes. The first argument is simply the AverageIfs formula itself, and the second is 2 meaning that only 2 decimal points will be shown as the results.